...G...... [ZBXNEXT-2225] one cast to "double" is enough

+++ src/libs/zbxsysinfo/win32/swap.c
@@ -106,7 +106,7 @@ int	SYSTEM_SWAP_SIZE(AGENT_REQUEST *request, AGENT_RESULT *result)
 	else if (0 == strcmp(mode, "free"))
 		SET_UI64_RESULT(result, real_swap_avail);
 	else if (0 == strcmp(mode, "pfree"))
-		SET_DBL_RESULT(result, ((double)real_swap_avail / (double)real_swap_total) * 100.0);
+		SET_DBL_RESULT(result, (real_swap_avail / (double)real_swap_total) * 100.0);
 	else if (0 == strcmp(mode, "used"))
 		SET_UI64_RESULT(result, real_swap_total - real_swap_avail);
 	else
